# Refactor Dockerfile and speed up build times via buildkit caches and layers