--- a/Dockerfile
+++ b/Dockerfile
@@ -242,11 +242,8 @@ COPY --chown=mastodon:mastodon --from=yarn-install /opt/mastodon /opt/mastodon
 ##########################################################################################
 # Assets layer
 ##########################################################################################
-#
-# NOTE: We force the platform to be the "native" build platform (e.g. linux/amd64 in GitHub Actions)
-# for this step, since web assets don't depend on "server-side" CPU architectures, and being able
-# to skip CPU emulation will significantly speed things up
-FROM --platform=$BUILDPLATFORM runtime-layer AS assets-precompile
+
+FROM runtime-layer AS assets-precompile
 
 RUN \
   # --mount=type=cache,target=/opt/mastodon/tmp/cache,uid=${UID},gid=${GID},id=assets-cache-${BUILDPLATFORM},sharing=locked \