--- a/config/environments/production.rb
+++ b/config/environments/production.rb
@@ -132,3 +132,5 @@ Rails.application.configure do
 
   config.x.otp_secret = ENV.fetch('OTP_SECRET')
 end
+
+# some silly file change in ruby