👋 Hello!

## Context

I noticed that in both the mastodon/mastodon and glitch-soc fork that builds can sometimes take hours, usually spending hours of time on work that could be effectively cached.

Since I work with CI/CD and have spent way too many cycles of my life in `Dockerfile` optimizations, I thought I would give it a shot on improving it for Mastodon, since it would make my own Mastodon server build times faster too 🎉  win/win.

I'm by no means a Ruby on Rails or `bundle` expert, so might not have aced the usage entirely (I did learn a bit from external resources on how to achieve the caching outcome though). 

## Changes

I fully recognize (and apologize for) the diff being very hard to read, given the amount of changes, so will do my best to outline changes at a high level here.

### Caching 

* Utilize the `RUN --mount=type=cache` to cache `apt-get update` and `apt-get install` operations
* Utilize the `RUN --mount=type=cache` to cache `bundle install` 
* Utilize the `RUN --mount=type=cache` to cache `yarn install` 
* Ensure apt caches aren't automatically cleaned by removing the apt script from the `base` layer

### Other

* Configure TimeZone via `TZ` (can have huge positive impact on performance!)
* Sorted packages alphabetically so it's easier to scan
* Slightly reformatted some Dockerfile statements to be easier to scan
* Expose some additional `ENV` as `ARG` for on-demand changing (useful during testing)
* Restructured the layers into multiple specific ones (e.g. `bundle` and `yarn`). The layering allow each to run in parallel rather than sequential
* Added comments to most statements in the `Dockerfile` with external references (docs) and reason for the statement to be there (hopefully useful for future maintainers) 

## Docker image sizes

This is the output of `docker build .` from (1) this PR (2) main and (3) DockerHub pull

```
jippi/mastodon              latest    b41bea14ec78   32 seconds ago      1.18GB
jippi/mastodon-main         latest    edfb04bd73e9   34 minutes ago      1.22GB
ghcr.io/mastodon/mastodon   latest    97e47539355f   28 hours ago        1.42GB
```

## Docker build times

Measured on an M1 Pro w/ 16GB RAM

```
this branch with no caches: 158.6s
this branch with no bundle caches, warm yarn: 104.2s
this branch with hot caches: 2.9s
```