@roobre 2023-07-07
Intruding here a bit since I also spent quite some time yesterday trying to optimize build time. How will `--mount=type=cache` improve build times on CI/CD systems, where each run occurs on a fresh instance with no traces from previous runs?