@renchap 2023-07-07
@jippi Thanks for your PR. We intend to re-work our Dockerfile soon, and most probably rewrite it pretty much entirely after 4.2.0 is released, as the build chain will change a lot (no need to have Node in the final image for example).

I will try to have a look at your PR in the coming days (weeks?), but you are not the first one to have attempted this, and improvements to the build process have been hard to merge until now.

> Intruding here a bit since I also spent quite some time yesterday trying to optimize build time. How will `--mount=type=cache` improve build times on CI/CD systems, where each run occurs on a fresh instance with no traces from previous runs?

If we use this in the Github action:
```
cache-from: type=gha
cache-to: type=gha,mode=max
```

Then Docker caches can be reused between actions.