@jippi 2023-07-07
@roobre as @renchap mentioned, https://github.com/moby/buildkit#github-actions-cache-experimental will ensure the layer caching is durable in CI :)