@jippi 2023-07-07
@renchap Ah okay, my changes should make the changes you propose *very* easy to achieve, since everything node is nicely isolated in an intermediate build target, so the final output layer can just use a different "source" than the one currently in place. 

The most "noisy" part of the PR is actually splitting things up so each layer can be more "specialized" _and_ run in parallel without stepping on each other. 

Without knowing the full details of your rewrite, I think my PR would make iterating and modifying it like you mention significantly easier 

I would be happy to help adjust those if you want to collaborate.