@renchap 2023-07-07
From what we are seeing in the builds, the most costly step (by far) is assets precompiling, and in particular running Webpack.

I am not sure why but it seems to be taking 2+ hours on CI.

I suspect that this is because we are building 2 archs on CI, and building `arm64` on an `amd64` worker is really not efficient.

If you want to have a further look at this feel free, you should be able to get all the logs from Github Actions.