@jippi 2023-07-07
@renchap need to figure out how to run the pipeline in my fork this weekend, I ended up caching some webpack and sprockets data as well, for the heck of it, which seemed to speed those up a bit when making trivially small changes - and making them entirely NOOP when no changes happened.

The CPU emulation is horribly slow, especially on the default small CPU allocations of the free tier. Stepping up one or two sizes have a significant improvement on the emulated CPU performance.

I do wonder if it wouldn't be possbile to build the assets only on amd64 and just copy it to the arm64 image - the asserts should be entirely identical to the browser, maybe something I can dig into optimizing if you would want that :)