@renchap 2023-07-07
> I do wonder if it wouldn't be possbile to build the assets only on amd64 and just copy it to the arm64 image - the asserts should be entirely identical to the browser, maybe something I can dig into optimizing if you would want that :)

Yes, the assets should be the same. We indeed should be able to run `COPY --from=…` for the assets, but then it means we have a different Dockerfile for each architecture, which I dont really like.

We could also build the 2 archs in //, but then you need to pull both of them and push them together on the registry, as we want the same tag but with 2 different arch.

If it helps, we can also consider having an ARM runner to build those images, but now we need to coordinate building on multiple runners and then doing a single push…