@renchap 2023-07-07
> you shouldn't need two different dockerfiles, no? just potentially a multi-stage build with the webpack step occurring in the build platform arch, and the final copying happening in the target platform arch (might be misunderstanding something, though)

Can you use multiple arch in the same Dockerfile? I did not knew this!