@jippi 2023-07-07
yep, this is exactly what I'm thinking - trying that out now :)