@davidlougheed 2023-07-07
@renchap here's an example I've worked on where the webpack build step happens on the native arch and the final stage happens on the target arch: https://github.com/bento-platform/bento_web/blob/master/Dockerfile