@jippi 2023-07-07
At $DayJob we just configured remote buildx connections for each arg, so buildx in CI job coordinated build steps across two native CPU architectures, but doesn't look like GitHub Actions makes that easy to do 😭