@renchap 2023-07-07
I think that you can do it using https://github.com/docker/setup-buildx-action, they have a section about configuring external builders.

But if the Webpack build is done only on the native platform, I suspect the build times will be much much better.