@roobre 2023-07-07
> cache-from: type=gha
> cache-to: type=gha,mode=max

> Then Docker caches can be reused between actions.

My understanding is that this works well for caching layers across runs, but I am not 100% sure about cache mounts. Has this been verified?

> But if the Webpack build is done only on the native platform, I suspect the build times will be much much better.

This is possible but would unfortunately require copy-paste into the Dockerfile. I was playing exactly with this yesterday, and was planning to give it a PoC PR tomorrow. I'd happily do so if you think it would be useful before the Dockerfile refactor for 4.2.0 you mentioned :)