@jippi 2023-07-07
Honestly, the best way to solve the slow speeds is to either pick a CI provider that provide native arm64/amd64 or increase the number of CPU cores on the builders to 4 or 8 for these steps. 

The hit of a cold cache will likely still run into "hours" of build time on small 2 core CI runners like GitHub provides for free. 

I tried for the heck of it to build the file on my Raspberry PI and it wildly outperformed my Apple M1 pro CPU (emulating amd64, on 8 cores)