@roobre 2023-07-08
> or increase the GitHub runner size to get increased number of CPU cores on the builders to 4 or 8 for this workload.

I don't think this will help a lot. Asset compilation is (for reasons unknown to me, I'm not a ruby ecosystem connoisseur) single-threaded, so at most you may be parallelizing some other steps but the bottleneck won't go away.

> pick a CI provider that provide native arm64/amd64

This would be the way, but will require additional effort and maintainer time.


I have a WIP patch that runs the asset build step on the native arch, at the cost of complicating the Dockerfile a bit. This brings down the time from ~1hour (on my workstation) to 11 minutes (also on my workstation), with fully clean buildx caches.

I'll keep testing this to see if it works, and if it does I'll open a PR with the patch. I think the mount cache changes would be very much additive to my changes, so there's more to gain :)