@jippi 2023-07-08
Amazing work with the Dockerfile refactor - I dabbled a bit with ARM native build last night, but doesn't look like GitHub Actions provide any native arm64 / aarch64 runners at all - or maybe I missed it? 
So that path would require an external vendor to be looped in (likely with associated cost?) 

If we have native runners, then its pretty easy to get buildx to do all the hard work of building on the right runner arch out of the box with little to no modifications to the Dockerfile itself